pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Production-oriented FastAPI starter template."
requires-python = ">=3.14"
dependencies = [
    "alembic==1.16.5",
    "arq==0.26.3",
    "bcrypt==4.0.1",
    "email-validator==2.3.0",
    "fastapi==0.128.5",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-http==1.38.0",
    "opentelemetry-instrumentation-fastapi==0.59b0",
    "opentelemetry-sdk==1.38.0",
    "passlib[bcrypt]==1.7.4",
    "prometheus-fastapi-instrumentator==7.1.0",
    "psycopg[binary]==3.2.13",
    "pydantic==2.12.5",
    "pydantic-settings==2.11.0",
    "python-dotenv==1.2.1",
    "python-jose[cryptography]==3.5.0",
    "python-json-logger==4.0.0",
    "python-multipart==0.0.20",
    "redis==5.3.1",
    "sentry-sdk[fastapi]==2.44.0",
    "sqlalchemy==2.0.46",
    "uvicorn[standard]==0.39.0",
]

[dependency-groups]
dev = [
    "fakeredis==2.32.1",
    "httpx==0.28.1",
    "hypothesis==6.141.1",
    "mypy==1.19.1",
    "pre-commit==4.3.0",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "ruff==0.15.0",
    "uv==0.9.11",
]

[tool.uv]
default-groups = ["dev"]

[tool.mutmut]
paths_to_mutate = ["app/calculations.py", "app/oauth2.py"]
tests_dir = [
    "tests/test_calculations.py",
    "tests/unit/test_oauth2_unit.py",
    "tests/property/test_calculations_property.py",
    "tests/property/test_auth_property.py",
]
do_not_mutate = ["*/__init__.py"]
also_copy = ["app"]
